Try to make "slow down" message that one could receive on rapid teleporting more informative to the user.
This message is seen on V2 if one attempts to quickly re-teleport from a source region where one had previously teleported to a non-neighbour and back within 15 secs.
The solution here is for the user to wait a short while.
This message can also be seen on any teleport protocol if one recieves multiple teleport attempts simultaneously.  Probably still useful here to help identify misbehaving scripts.