Fix runprebuild.bat on Windows 7 systems and later
MS changed Windows 7 "reg query" command to spit out fewer lines in its preamble
Annoyingly there is no option to print raw values and not this preamble.
So now using the Windows find command.
But this has the side effect of stopping this working with Cygwin as it uses its own GNU find command.
Might fix in the future but not simple since Windows find itself is in different locations